Consolidate RapidsConf entry groups
